requirements/requirements.txt
certifi==2017.4.17
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.2
django-allauth==0.32.0
django-cors-headers==2.1.0
django-rest-auth==0.9.1
djangorestframework==3.6.3
html5lib==1.0b10
httpie==0.9.9
httplib2==0.10.3
idna==2.5
oauthlib==2.0.2
olefile==0.44
Pillow==4.1.1
Pygments==2.2.0
PyPDF2==1.26.0
python3-openid==3.1.0
pytz==2017.2
reportlab==3.4.0
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.21.1
webencodings==0.5.1
xhtml2pdf==0.2b1
